Уведомления вынесены в отдельный класс
