+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/di/AppContainer.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import androidx.work.WorkManager
 import com.alexanderpodkopaev.androidacademyproject.data.MoviesDatabase
 import com.alexanderpodkopaev.androidacademyproject.data.RetrofitModule
+import com.alexanderpodkopaev.androidacademyproject.notifications.MoviesNotification
 import com.alexanderpodkopaev.androidacademyproject.repo.*
 import com.alexanderpodkopaev.androidacademyproject.service.MoviesSyncSchedule
 
@@ -15,4 +16,5 @@ class AppContainer(context: Context) {
     val moviesRepository: MoviesRepository = MoviesRepoImpl(RetrofitModule.moviesApi, dbDataSource)
     val actorsRepository: ActorsRepository = ActorsRepoImpl(RetrofitModule.moviesApi, dbDataSource)
     val moviesSyncSchedule: MoviesSyncSchedule = MoviesSyncSchedule(workManager)
+    val moviesNotification: MoviesNotification = MoviesNotification(context)
 }
