+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/ui/MainActivity.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
+import com.alexanderpodkopaev.androidacademyproject.MyApp
 import com.alexanderpodkopaev.androidacademyproject.R
 
 class MainActivity : AppCompatActivity() {
@@ -23,18 +24,14 @@ class MainActivity : AppCompatActivity() {
 
     override fun onNewIntent(intent: Intent?) {
         super.onNewIntent(intent)
-        Log.d("MyWorker", "Start notification")
         if (intent != null) {
             handleIntent(intent)
         }
     }
 
     private fun handleIntent(intent: Intent) {
-        Log.d("MyWorker", "in handle")
         when (intent.action) {
             Intent.ACTION_VIEW -> {
-                Log.d("MyWorker", "in intent")
-
                 val id = intent.data?.lastPathSegment?.toIntOrNull()
                 if (id != null) {
                     openMovie(id)
@@ -44,11 +41,11 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun openMovie(id: Int) {
-        Log.d("MyWorker", "open fragment")
         supportFragmentManager.beginTransaction()
             .replace(R.id.flFragment, FragmentMoviesDetails.newInstance(id))
             .addToBackStack(null)
             .commit()
+        MyApp.container.moviesNotification.dismissNotification(id)
     }
 
 }
